requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-cloud-firestore==2.19.0
httpx==0.27.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
